Set new type

+++ src/eyreish/context.rs
@@ -134,7 +134,7 @@ where
         self.error.severity()
     }
 
-    fn line_offset(&self) -> Option<u32> {
+    fn line_offset(&self) -> Option<usize> {
         self.error.line_offset()
     }
 
@@ -171,7 +171,7 @@ where
         unsafe { ErrorImpl::diagnostic(self.error.inner.by_ref()).severity() }
     }
 
-    fn line_offset(&self) -> Option<u32> {
+    fn line_offset(&self) -> Option<usize> {
         unsafe { ErrorImpl::diagnostic(self.error.inner.by_ref()).line_offset() }
     }
